Comply with startMarks Configuration Option
In the configuration object, you can include a list
`config$behavior$startMarks` of marks that will be used as the start of
the assay. These should be listed in order of priority because the first
one in the list that is found in a log file is used as that log's start
mark. If no match is found, any user-specified default is used. If no
default is specified, the user is prompted to select one.